refactor: create a common Table component and use it on approved & rejected users lists
