+++ src/components/Dashboard/RejectedUsers/index.jsx
@@ -1,59 +1,13 @@
-import { useEffect, useState } from 'react';
-
-import { useAppDispatch, useAppSelector } from 'hooks';
 import { getRejectedList } from 'store/admin/thunk';
-import columns from 'objects/Users';
-import { Loader } from 'shared/components';
-import {
-  Input, Table,
-} from 'components/AntDesign';
+import { UsersTable } from 'shared/components';
 
 function RejectedUsers() {
-  const [value, setValue] = useState('');
-  const [dataSource, setDataSource] = useState([]);
-  const dispatch = useAppDispatch();
-  const [rejectedList, isLoading] = useAppSelector((state) => [
-    state.admin.rejectedList,
-    state.admin.isLoading,
-  ]);
-
-  useEffect(() => {
-    dispatch(getRejectedList());
-  }, []);
-
-  useEffect(() => {
-    setDataSource(rejectedList);
-  }, [rejectedList]);
-
-  const FilterByNameInput = (
-    <Input
-      placeholder="Search user by Name"
-      allowClear
-      value={value}
-      onChange={(e) => {
-        const currValue = e.target.value;
-        setValue(currValue);
-        const filteredData = rejectedList.filter((entry) => entry.username.toLowerCase()
-          .includes(currValue));
-        setDataSource(filteredData);
-      }}
+  return (
+    <UsersTable
+      listToDisplay="rejectedList"
+      thunkFunction={getRejectedList}
     />
   );
-
-  return (
-    !isLoading ? (
-      <div className="sort">
-        <div className="search">{FilterByNameInput}</div>
-        <div className="table">
-          <Table
-            columns={columns}
-            dataSource={dataSource}
-            pagination={{ pageSize: 5 }}
-            scroll={{ x: 500 }}
-          />
-        </div>
-      </div>
-    ) : <Loader />);
 }
 
 export default RejectedUsers;
