+++ src/shared/components/index.js
@@ -1,2 +1,2 @@
-// eslint-disable-next-line import/prefer-default-export
 export { default as Loader } from './Spinner';
+export { default as UsersTable } from './Table';
